Added to description.. Easy to customize
I added Easy to Customize because it is but I can omit that if you want. If that is too long. I don't think it is by glancing at the other framework descriptions and 2 lines given for the description. 

My thinking was there are 4 points to point out about this Codestar's Framework. 

1. 5 in 1 - Metaboxes, Taxonomies, Customizer, Theme Options, etc. 
2. It's easy to customize
3. It's light weight
4. It's for Themes and Plugins